pyproject.toml
[tool.poetry]
name = "fastapiproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "2.0.40"
fastapi = "0.115.12"
pydantic = {extras = ["email"], version = "^2.11.4"}
pydantic-settings = "2.9.1"
python-multipart = "0.0.20"
alembic = "1.15.2"
asyncpg = "0.30.0"
greenlet = "3.2.2"
psycopg2 = "2.9.10"
minio = "7.2.15"
uvicorn = "^0.34.2"
passlib = "^1.7.4"
python-jose = "^3.4.0"
pycryptodome = "^3.22.0"
bcrypt = "^4.3.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
